requirements.txt
alembic==1.9.4
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.2
packaging==23.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
swagger-ui-bundle==0.0.9
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
